Improving the way as the crop tool work to maximize the performance. Added too some language constants on the crop tool.

+++ htdocs/modules/system/admin/images/main.php
@@ -538,7 +538,7 @@ function imanager_listimg($imgcat_id,$start=0) {
 			$params .= '&image_id='.$images[$i]->getVar('image_id');
 			$params .= '&uniq='.$uniq;
 			$params .= '&crop_script='.ICMS_URL.'/modules/system/admin/images/crop_image.php';
-			$arrimg[$i]['crop_link'] = 'window.open(\''.ICMS_LIBRARIES_URL.'/image-crop/image-crop.php'.$params.'\',\'imageWin\',\'width='.($arrimg[$i]['width']+227).',height='.(($arrimg[$i]['height']<=400)?400:$arrimg[$i]['height']+30).',resizable=yes\'); return false;';
+			$arrimg[$i]['crop_link'] = 'window.open(\''.ICMS_LIBRARIES_URL.'/image-crop/image-crop.php'.$params.'\',\'imageWin\',\'width='.($arrimg[$i]['width']+227).',height='.(($arrimg[$i]['height']<=500)?500:$arrimg[$i]['height']+30).',resizable=yes\'); return false;';
 		}else{
 			$arrimg[$i]['hasextra_link'] = 0;
 		}
@@ -1208,31 +1208,60 @@ function imanager_cropimg() {
 	$url = (substr($categ_url,-1) != '/')?$categ_url.'/':$categ_url;
 	$path = (substr($categ_path,-1) != '/')?$categ_path.'/':$categ_path;
 	
-	if ($imagecategory->getVar('imgcat_storetype') == 'db') {
-		$fp = @fopen($path.'crop_'.$uniq.'_'.$newimg, 'rb');
-		$fbinary = @fread($fp, filesize($path.'crop_'.$uniq.'_'.$newimg));
-		@fclose($fp);
-		$image->setVar('image_body', $fbinary, true);
-		if (!$image_handler->insert($image)) { //Falha ao salvar db, apagar imagem crop
-			$msg = _MD_FAILEDIT;
-			@unlink($path.'crop_'.$uniq.'_'.$newimg);
-		}else{ //Db salvo, apagar img original, copiar imagem crop para imagem nova, apagar imagem crop
-			@unlink($path.'crop_'.$uniq.'_'.$newimg);
-			$msg = _MD_AM_DBUPDATED;
-		}
-	}else{
-		if (unlink($path.$oldimg)){
-			if (copy($path.'crop_'.$uniq.'_'.$newimg,$path.$newimg)){
-				if (unlink($path.'crop_'.$uniq.'_'.$newimg)){
-					$msg = _MD_AM_DBUPDATED;
+	if ($overwrite){
+		if ($imagecategory->getVar('imgcat_storetype') == 'db') {
+			$fp = @fopen($path.'crop_'.$uniq.'_'.$newimg, 'rb');
+			$fbinary = @fread($fp, filesize($path.'crop_'.$uniq.'_'.$newimg));
+			@fclose($fp);
+			$image->setVar('image_body', $fbinary, true);
+			if (!$image_handler->insert($image)) { //Falha ao salvar db, apagar imagem crop
+				$msg = _MD_FAILEDIT;
+				@unlink($path.'crop_'.$uniq.'_'.$newimg);
+			}else{ //Db salvo, apagar img original, copiar imagem crop para imagem nova, apagar imagem crop
+				@unlink($path.'crop_'.$uniq.'_'.$newimg);
+				$msg = _MD_AM_DBUPDATED;
+			}
+		}else{
+			if (unlink($path.$oldimg)){
+				if (copy($path.'crop_'.$uniq.'_'.$newimg,$path.$newimg)){
+					if (unlink($path.'crop_'.$uniq.'_'.$newimg)){
+						$msg = _MD_AM_DBUPDATED;
+					}else{
+						$msg = sprintf(_MD_FAILUNLINK,$path.'crop_'.$uniq.'_'.$newimg);
+					}
 				}else{
-					$msg = sprintf(_MD_FAILUNLINK,$path.'crop_'.$uniq.'_'.$newimg);
+					$msg = _MD_FAILEDIT;
 				}
 			}else{
 				$msg = _MD_FAILEDIT;
 			}
+		}
+	}else{
+		$ext = substr($image->getVar('image_name'),strlen($image->getVar('image_name'))-3,3);
+		$imgname = 'img'.icms_random_str(12).'.'.$ext;
+		$newimg =& $image_handler->create();
+		$newimg->setVar('image_name', $imgname);
+		$newimg->setVar('image_nicename', $_GET['image_nicename']);
+		$newimg->setVar('image_mimetype', $image->getVar('image_mimetype'));
+		$newimg->setVar('image_created', time());
+		$newimg->setVar('image_display', $_GET['image_display']);
+		$newimg->setVar('image_weight', $_GET['image_weight']);
+		$newimg->setVar('imgcat_id', $image->getVar('imgcat_id'));
+		if ($imagecategory->getVar('imgcat_storetype') == 'db') {
+			$fp = @fopen($path.'crop_'.$uniq.'_'.$oldimg, 'rb');
+			$fbinary = @fread($fp, filesize($path.'crop_'.$uniq.'_'.$oldimg));
+			@fclose($fp);
+			$newimg->setVar('image_body', $fbinary, true);
+			@unlink($path.'crop_'.$uniq.'_'.$oldimg);
 		}else{
-			$msg = _MD_FAILEDIT;
+			if (copy($path.'crop_'.$uniq.'_'.$oldimg,$path.$imgname)){
+				@unlink($path.'crop_'.$uniq.'_'.$oldimg);
+			}
+		}
+		if (!$image_handler->insert($newimg)) {
+			$msg = sprintf(_FAILSAVEIMG, $newimg->getVar('image_nicename'));
+		}else{
+			$msg = _MD_AM_DBUPDATED;
 		}
 	}
